issue #3: Adding new tasks does not change the selection of the current task.
So no stopping of a running task happens. When no activity is current (no task started), then the message dialog does not appear.
The special situation, that the repository is empty and the first task is added, is treated different: In this case the first task introduced is automatically preselected in the selector.